Rename parser to edit
